requirements.txt
fastapi
uvicorn
python-dotenv
huggingface-hub
pydantic
supabase
sentence-transformers
pypdf2
python-multipart
yt_dlp
llama-index
ollama
torch 
torchvision
opencv-python
cohere
numpy<2
matplotlib
scikit-learn
datasets
ragas
pandas
langchain-groq